+++ src/semble/installer/__init__.py
@@ -1,0 +1,3 @@
+from semble.installer.installer import run
+
+__all__ = ["run"]
